pyproject.toml
[project]
name = "meshque"
version = "0.1.0"
description = "E-commerce platform built with Django"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "pillow>=12.0.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]
